feat: moved css styles to css folder
