+++ packages/components/src/components/include/include.styles.ts
@@ -1,5 +1,5 @@
 import { css } from 'lit';
-import componentStyles from '../../styles/component.styles';
+import componentStyles from '../css/component.styles';
 
 export default css`
   ${componentStyles}
